Add a load_status callback hook that is invoked at the end of chrome LoadApp.
This callback will be used to communicate the load status (success
or error) in the place of the start_module SRPC call, so that we can
deprecate the "start_module" call later. Blocking on "start_module"
is kept for now.

Next step is to use it in chrome like so:
https://codereview.chromium.org/1090233003/

When the hook is set, assume that we want to reap logs on error,
so that we can remove the SRPC "log" command as well.

BUG= https://code.google.com/p/chromium/issues/detail?id=459250

Review URL: https://codereview.chromium.org/1089323006