Implement a standard ELF loader in untrusted code, to be used by glibc
In the new glibc port (as in non-NaCl ELF generally), the dynamic linker
is a standard ET_DYN object rather than an ET_EXEC nexe, and expects to
be loaded via the PT_INTERP mechanism.  Instead of turning the dynamic
linker into the main nexe, a small ELF chain-loader is used as the main
nexe, and it loads the executable and dynamic linker following the
normal ELF procedure.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3068
TEST= trybots & new tests run_elf_loader_*_test_irt
R=dschuff@chromium.org, phosek@chromium.org

Review URL: https://codereview.chromium.org/1090143002