Update revision of arm sysroot image in arm_trusted.json
This includes the latest updates to ubuntu/precise.  The
archive was created using:
 trusted-toolchain-creator.armhf.precise.sh BuildJail sysroot_arm_trusted_precise.tg
and uploaded using:
 gsutil cp -a public-read sysroot_arm_trusted_precise.tg gs://nativeclient-archive2/toolchain/70d11372070d5f64335b47a40b76eb478f42e2f2/

This change was generated by editing the json file directly
and then running:
./build/package_version/package_version.py recalcrevisions

This change is simply to being the sysroot image up-to-date.
They successful the next change will be to rev the
qemu version from 1.0.1 to 2.3.0.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4157

Review URL: https://codereview.chromium.org/1118543002