Enable suppression of elf_ldr build in run.py
Run.py now considers elf_ldr.nexe a prerequisite along with irt_core.nexe
In non-glibc arm SCons test runs, elf_ldr.nexe is not built when run.py
is invoked. So allow suppression of its automatic build in the same way
we do with the IRT, and use the suppression in the same way for the
run_py_test.

BUG=run_py tests failing on arm bots
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1122913002